+++ src/server/API/Startup.cs
@@ -1,4 +1,5 @@
 using FluentPOS.Modules.Catalog;
+using FluentPOS.Modules.Catalog.Extensions;
 using FluentPOS.Modules.Identity.Extensions;
 using FluentPOS.Modules.People.Extensions;
 using FluentPOS.Shared.Core.Extensions;
